Fix Tram and Pauline file locations
