--- Animations.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d5334cc15a6008848a27a1314dca7d3f
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
