--- Assets/Resources/Sprites/Games/Tram&Pauline/Tram&Pauline.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 779f3d97d97807a41b869aac547fdd96
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
