+++ Assets/Resources/Sprites/Games/Tram&Pauline/Tram&Pauline_Colors.png
