+++ Assets/Resources/Sprites/Games/Tram&Pauline/Tram&Pauline_Curtains.png
