+++ Assets/Resources/Sprites/Games/Tram&Pauline/Tram&Pauline_SpriteSheet.png.meta
@@ -4,7 +4,7 @@ TextureImporter:
   internalIDToNameTable:
   - first:
       213: 1116226217280006615
-    second: Tram-&-Pauline-Sprite-Sheet 1_0
+    second: Tram-&-Pauline-Sprite-Sheet_0
   - first:
       213: 7574086205227880532
     second: Tram-&-Pauline-Sprite-Sheet 1_1
@@ -211,7 +211,7 @@ TextureImporter:
     serializedVersion: 2
     sprites:
     - serializedVersion: 2
-      name: Tram-&-Pauline-Sprite-Sheet 1_0
+      name: Tram-&-Pauline-Sprite-Sheet_0
       rect:
         serializedVersion: 2
         x: 3
@@ -513,7 +513,7 @@ TextureImporter:
         width: 45
         height: 60
       alignment: 0
-      pivot: {x: 0, y: 0}
+      pivot: {x: 0.5, y: 0.5}
       border: {x: 0, y: 0, z: 0, w: 0}
       outline: []
       physicsShape: []
@@ -534,7 +534,7 @@ TextureImporter:
         width: 45
         height: 60
       alignment: 0
-      pivot: {x: 0, y: 0}
+      pivot: {x: 0.5, y: 0.5}
       border: {x: 0, y: 0, z: 0, w: 0}
       outline: []
       physicsShape: []
