--- Assets/Resources/Sprites/Games/Tram&Pauline/TramAndPauline.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8850749af64f82b44a728e2c273157e1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
