--- Assets/Resources/Sprites/Trampolines.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 47beef580fcf1514099d4f00edc3988c
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
