+++ Assets/Scripts/Games/Tram&Pauline/Curtains.cs
