--- Effects.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: dd4cf4b851b908449914b11be594600f
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
